+++ problem-service/src/main/kotlin/com/sevity/problemservice/domain/ProblemRepository.kt
@@ -3,4 +3,4 @@ package com.sevity.problemservice.domain
 import com.sevity.problemservice.domain.Problem
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ProblemRepository : JpaRepository<Problem, Long>
+interface ProblemRepository : JpaRepository<Problem, Int>
